touriquest-backend/services/auth-service/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1

# Authentication and security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7

# HTTP client for OAuth
httpx==0.25.2
aiohttp==3.9.1

# Email dependencies
aiosmtplib==2.0.2
jinja2==3.1.2

# Redis for caching and sessions
redis==5.0.1
aioredis==2.0.1

# Environment and configuration
python-dotenv==1.0.0
pydantic[email]==2.5.0
pydantic-settings==2.0.3

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3